Separate var generation into separate VarGen
