+++ linker/shared/src/main/scala/org/scalajs/linker/backend/emitter/ClassEmitter.scala
@@ -39,6 +39,7 @@ private[emitter] final class ClassEmitter(jsGen: JSGen) {
   import functionEmitter._
   import jsGen._
   import nameGen._
+  import varGen._
 
   def buildClass(tree: LinkedClass, ctor: WithGlobals[js.Tree],
       memberDefs: List[WithGlobals[js.Tree]], exportedDefs: WithGlobals[js.Tree])(
